requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
alembic==1.13.1
    # via -r requirements.in
amqp==5.3.1
    # via kombu
apscheduler==3.11.0
    # via flask-apscheduler
bidict==0.23.1
    # via python-socketio
billiard==4.2.1
    # via celery
blinker==1.9.0
    # via flask
celery==5.5.3
    # via -r requirements.in
click==8.2.1
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   flask
click-didyoumean==0.3.1
    # via celery
click-plugins==1.1.1.2
    # via celery
click-repl==0.3.0
    # via celery
colorama==0.4.6
    # via
    #   click
    #   pytest
dnspython==2.7.0
    # via eventlet
eventlet==0.40.2
    # via -r requirements.in
flask==3.1.1
    # via
    #   -r requirements.in
    #   flask-admin
    #   flask-apscheduler
    #   flask-socketio
    #   flask-sqlalchemy
flask-admin==1.6.1
    # via -r requirements.in
flask-apscheduler==1.12.3
    # via -r requirements.in
flask-socketio==5.5.1
    # via -r requirements.in
flask-sqlalchemy==3.1.1
    # via -r requirements.in
greenlet==3.2.3
    # via
    #   eventlet
    #   sqlalchemy
gunicorn==23.0.0
    # via -r requirements.in
h11==0.16.0
    # via wsproto
iniconfig==2.1.0
    # via pytest
itsdangerous==2.2.0
    # via flask
jinja2==3.1.6
    # via flask
kombu==5.5.4
    # via celery
mako==1.3.10
    # via alembic
markupsafe==3.0.2
    # via
    #   flask
    #   jinja2
    #   mako
    #   werkzeug
    #   wtforms
mysql-connector-python==9.3.0
    # via -r requirements.in
networkx==3.5
    # via -r requirements.in
packaging==25.0
    # via
    #   gunicorn
    #   kombu
    #   pytest
pluggy==1.6.0
    # via pytest
prompt-toolkit==3.0.52
    # via click-repl
pygments==2.19.2
    # via pytest
pytest==8.4.1
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via
    #   celery
    #   flask-apscheduler
python-dotenv==1.1.1
    # via -r requirements.in
python-engineio==4.12.2
    # via python-socketio
python-socketio==5.13.0
    # via flask-socketio
pytz==2025.2
    # via -r requirements.in
pywin32==310 ; sys_platform == "win32"
    # via -r requirements.in
redis==6.4.0
    # via -r requirements.in
simple-websocket==1.1.0
    # via python-engineio
six==1.17.0
    # via python-dateutil
sqlalchemy==2.0.27
    # via
    #   alembic
    #   flask-sqlalchemy
tabulate==0.9.0
    # via -r requirements.in
typing-extensions==4.14.1
    # via
    #   alembic
    #   sqlalchemy
tzdata==2025.2
    # via
    #   kombu
    #   tzlocal
tzlocal==5.3.1
    # via apscheduler
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.13
    # via prompt-toolkit
werkzeug==3.1.3
    # via flask
wsproto==1.2.0
    # via simple-websocket
wtforms==3.2.1
    # via flask-admin
